+++ lib/node-pre-gyp.js
@@ -10,13 +10,6 @@ module.exports = exports;
  * Module dependencies.
  */
 
-// load mocking control function for accessing s3 via https. the function is a noop always returning
-// false if not mocking.
-exports.mockS3Http = require('./util/s3_setup').get_mockS3Http();
-exports.mockS3Http('on');
-const mocking = exports.mockS3Http('get');
-
-
 const fs = require('fs');
 const path = require('path');
 const nopt = require('nopt');
@@ -41,10 +34,6 @@ const cli_commands = [
   'configure'
 ];
 const aliases = {};
-
-if (mocking) {
-  log.warn(`mocking s3 to ${process.env.node_pre_gyp_mock_s3}`);
-}
 
 // this is a getter to avoid circular reference warnings with node v14.
 Object.defineProperty(exports, 'find', {
